Create Configuration::getDriverOptions() to create the client
